ref(feedback): split up SentryUserFeedbackForm.swift (#4726)
